Replace gethostbyname() with getaddrinfo() and improve logging at connect().
This changeset replaces gethostbyname() with getaddrinfo().  More work is
however needed to refactor how getaddrinfo() and connect() interacts.  At
the moment these calls are in different files and functions, so we only
try the first record returned.

Also, due to recent problems with one of the dyndns.org checkip servers the
logging at connect has been improved quite a bit.  Now inadyn displays the
acutal name and IP address of the checkip server it connects to.  This will
help rule out problems with inadyn and hopefully flush out problems with DNS
caching in the Os, or intermediate DNS proxys.  For more info on this, see
GitHub issue #3.