When starting: always use cache file, if it exists, or seed with DNS lookup.
On embedded systems, without a valid system battery backed up RTC, the system
time may skip at around the same time inadyn starts, which may or may not be
on the "correct" side of the time skip when the inadyn.conf file was created.

So, when inadyn (re)starts with or without a cache file we always try to use
that cache file to seed each configured alias' name:ip pair.  If  the cache
file doesn't exist we try to use DNS lookup to seed instead.

However, if we are in fact in a startup phase we may not yet have a full
access to the Internet yet.  (DHCP lease not yet given, routes not setup,
or ports on a switch not yet in forwarding.)  That we still do not have
support for, maybe we should add a connectivity check or postponed start
option?  I.e., to not proceed until Internet connectivity is secured, or
a startup-delay has been reached.

a