Stop handling SIGQUIT and SIGALRM
1) No longer use SIGQUIT to stop inadyn.
SIGQUIT is intended to initiate a core dump for diagnostic purposes.

2) No longer use SIGALRM to stop inadyn.
POSIX allows the implementation of sleep(3) and usleep(3) [1] to be
implemented using SIGALRM.  Blocking or ignoring the signal or otherwise
changing the time when SIGALRM is scheduled to be generated causes
unspecified behaviour; i.e. platform specific behaviour not defined by
POSIX.  On Linux, Glibc's implementation of usleep(3) doesn't use
SIGALRM, but on other platforms it may [2].

3) Block exactly the handled signals while the handler is running.
Not necessary, but just means the handler isn't called recursively if
two different signals are deliver simultaneously.

References:
[1] POSIX.1-2004, usleep(3) definition
    http://pubs.opengroup.org/onlinepubs/009695399/functions/usleep.html

[2] GNU C Library Reference Manual, 10.24 usleep
    http://www.gnu.org/software/gnulib/manual/html_node/usleep.html

Signed-off-by: Mike Fleetwood <mike.fleetwood@googlemail.com>