Move clearing of CMD_FORCED_UPDATE in DDNS client main loop
Scenario:
Inadyn is in the middle of an update check, in the call chain starting
with dyn_dns_update_ip(), when SIGUSR1 is received.

Example sequence of the previous code from that point onwards:
1) An OS call such as recv() is interrupted by the signal, returning an
   error.
2) The signal handler sets ctx->cmd = CMD_FORCED_UPDATE.
3) dyn_dns_update_ip() eventually returns an error.
4) DDNS client main loop clears ctx->cmd to NO_CMD before
   dyn_dns_wait_for_cmd().
Outcome is that forced update is NOT performed.

Fix by clearing ctx->cmd to NO_CMD immediately after force_addr_update
flag is set.  This is after all the only step which needs performing to
initiate a forced update.

Example sequence of the new code from the same point onwards:
[(1) to (3) as before]
4) DDNS client main loop calls dyn_dns_wait_for_cmd()
5) dyn_dns_wait_for_cmd() returns immediately because ctx->cmd is set to
   CMD_FORCED UPDATE.
6) FORCED_UPDATE log message is emitted and force_addr_update flag set.
Outcome is that the forced update is performed.

Signed-off-by: Mike Fleetwood <mike.fleetwood@googlemail.com>