Fix issue #35 and #48 -- Add support for one cache file per DNS alias
This commit adds support for multiple cache files, one for each DNS
entry.  As such it helps a great deal to keep track of responses from
different DDNS servers.  It should also fix issue #48, since upon
restart inadyn would fail to locate a cache file for a new DNS entry.

Also, a lot of cleanup and simplification for the support for the
various DDNS providers.  The next step should be to relocate each
provider into a separate "plugin", or similar.

Bumping version to 1.99.7-beta1, which is really just another one
down the road to 2.0 :)

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>