SSL mitigation fixes for POODLE, use SSL_MODE_SEND_FALLBACK_SCSV if available.
The intention of this commit is to prevent unpatched servers
from being misused.  We do this by disabling SSLv2 and SSLv3
when Inadyn is built with OpenSSL.  With GnuTLS nothing have
to be done since it already defaults to TLS.

For more details, see:

    https://www.openssl.org/news/secadv_20141015.txt
    http://marc.info/?l=openssl-dev&m=141333049205629&w=2
    https://www.openssl.org/~bodo/ssl-poodle.pdf

Also, use sizeof(buf) instead of hard coding the size of buf.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>