Add append-myip setting for custom@ DDNS providers
This patch adds support for a `append-myip` setting to `custom@` DDNS
providers.  Effectively it causes Inadyn to append the current IP
address, instead of the hostname alias, to the GET update URL sent to
the server.

README and man pages updated accordingly.

Also, the Markdown syntax in the README has been updated to handle
generic Markdown better, not just the simplified GitHub variant.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>