Remove generated files and add check for missing GIT submodules
This patch removes all files generated by the GNU configure and build
system, autoconf, autoheader and automake.  This does not affect the
disitrbuted tarballs, only the source code stored in GIT.

Also, add support for checking that the GIT submodules are OK, because
the GitHub ZIP files do not properly include the contents of submodules.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>