Add actual permissions check to os_check_perms()
This fix actually checks the file system permissions for the current
user.  With the recently fixed --drop-privs support we can now fully
detect that we cannot update the cache files, which is critical for
correct operation.

Needs further refactoring to get access to cache files when running
in --drop-privs mode.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>