Support for multiple users of the same provider, idea from Valery Frolov
This feature is inspired by a patch sent by Valery Frolov.  It enables
Inadyn to use the same provider for multiple users, very useful if you
act as a gateway for multiple downstream users.

    provider default@no-ip.com:1 {
        username    = ian
        password    = secret
        alias       = flemming.no-ip.com
	}

    provider default@no-ip.com:2 {
        username    = james
        password    = bond
        alias       = spectre.no-ip.com
    }

Thank you so much Valery for taking the time to explain the use-case to
me in such detail.  The resulting code is not compatible with the code
from your patch, but I hope it is in par with your request.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>