Fix #67: Use GnuTLS native API for HTTPS
This patch fixes both issue #67 (use native GnuTLS API for HTTPS) and
parts of issue #61 (certificate validateion), as well as the GnuTLS part
of issue #122 (segfault after SSL refactor).

The certificate validation and the GnuTLS support is blatantly stolen
from the following public domain example.

http://www.hep.by/gnu/gnutls/Simple-client-example-with-X_002e509-certificate-support.html

Parts of the code may need to be brought up to the correct API level,
and possibly synchronized with the GnuTLS version in the latest stable
Debian, but that's for another day.  Current GnuTLS support is now on
parity with Ubuntu 15.10.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>