Support external checkip-command in inadyn.conf
This patch adds support for calling an external shell script or command
to fetch the IP address.  Works for any type of DDNS provider.

The following environemnt variables are set:

- INADYN_PROVIDER: DDNS provider name, email@ddns-service.tld
- INADYN_USER    : User's name

The syntax is fairly straight forward:

    checkip-command = /sbin/ifconfig eth0 | grep 'inet addr'

The first occurrence matching an IPv4 or IPv6 address from the command's
output will be used by Inadyn.

Signed-off-by: Valery Frolov <fvgenn@gmail.com>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>