Add -I,--ident=NAME option for syslog+pidfile name
This patch adds support for the command line --ident=NAME, or -I NAME,
option.  It can be used to change the syslog and pidfile name of Inadyn
when running in multi-daemon setups, or if a user simply wants to mask
the identity of the DDNS client.  The client ID string sent over HTTP
is however not affected by this change, so the server will still know.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>