Fix #147: Refactor 5a6c117 to follow ssl option
This patch refactors 5a6c117 according to the discussion in #147, with a
little twist.  The checkip-ssl setting now defaults to follow the ssl
option, except when a custom checkip-server is configured by the user.

As of this refactor the built-in, hardcoded, knowledge of what providers
support HTTPS for their checkip-servers is replaced with updated
docs.  (The intention is to change default to SSL=true for Inadyn.)

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>