Refactor server connect(), fixes leaky abstraction
This patch is a refactor of how we connect() to the checkip and update
servers.  Previously the IP layer returned the address of only one of
the DNS records returned for the checkip or update servers.  Usually a
DDNS provider returns multiple IP records for the same checkip server,
this because the load on each one is quite high.  With the changes in
this patch we now try to connect(), with a lower timeout than before,
and on failure immediately try to connect to the next IP in the DNS
record.

Also, this fixes the annoying leaky abstraction between tcp.c and ip.c,
no need now to access internal IP structures from the TCP layer.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>