Use paths from configure in .conf and cache dir
This patch fixes the long standing issue with hard coded default values
for the .conf file and cache dir.  Instead of relying on the /etc and
/var prefixes this patch uses the configure script's SYSCONFDIR and
LOCALSTATEDIR, as provided by the user.

Also, inadyn now honors the --ident=NAME in the configuration file,
cache directory and PID file.  E.g., if the user wants to start two
independend inadyn processes, one called 'foo' and the other 'bar':

    inadyn --ident=foo
       => /etc/foo.conf
          /var/cache/foo/*.cache
	  /run/foo.pid

    inadyn --ident=bar
       => /etc/bar.conf
          /var/cache/bar/*.cache
	  /run/bar.pid

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>