Ignore premature termination in GnuTLS.
When receiving a reply using GnuTLS, gnutls_record_recv() may return
GNUTLS_E_PREMATURE_TERMINATION when the server closes the TCP
connection without a prior TLS shutdown alert. That probably indicates
a bug in the server's TLS handling.

Assuming the error is not really fatal, and because OpenSSL seems to
ignore it as well, check for and ignore this error code in the GnuTLS
implementation.

Also adds some log messages for GnuTLS connection failures.