Allow inadyn to run despite cache dir and PID file cannot be written
The mechanism for caching the IP address sent to each DDNS provider is
only to prevent too frequent, and unnecessary, updates.  If the system
does not allow us to write to the cache directory we carry on anyway.

The same goes for the PID file.  It is serves two purposes;

  1. To find the PID to send a signal to, consider the case where
     you run multiple inadyn clients with different identity, one
     per upstream interface, if you have multiple.
  2. To acknowledge a SIGHUP to any process/user monitoring inadyn

GitHub issue #284

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>