Refactor, simplify and possibly fix OpenSSL backend
This patch is a refactor of 110bde8, which added a recursive call to
ssl_send() and ssl_recv(), respectively, in each backend function.

The HTTP header and body are received in the same call, so no need
for a second call to SSL_read(), or SSL_write(), but we must make
sure to retry -- with the same buffer/offset -- as long as OpenSSL
gives us SSL_ERROR_WANT_WRITE or SSL_ERROR_WANT_READ.  It is the
EAGAIN equivalent.

Also, improve debug message a bit bit also logging max buffer len.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>