Initial unit test framework, requires external .conf files
In-a-Dyn is client for many different public DDNS services, these
require an account with credentials to test.  The test framwork
I'm adding here relies on my own personal credentials and those
that other users of the project donate.

The plan is to use GitHub Actions to run the actual tests and to
store all account secrets safely at GitHub.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>