Comment out logging of base-64 encoded password
Ever since we added, and then later defaulted to, HTTPS for all In-a-Dyn
communications with the outside world this log message has been somewhat
of a pain.  It's useful for debugging problems with encoding, which we
did have previously, but now it's just a constant security nightmare
when users post debug logs.  Let's disable it for now, maybe add another
command line switch for it in the future, if requested.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>