Ignore SIGPIPE so we can handle EPIPE from read() and SSL_read()
While testing #365 (memory leak in OpenSSL), I ran into a case when
SSL_read() failed and Inadyn exited on unhandled SIGPIPE.  Turns out
this happens when the remote end closes the connection while we are
waiting for a response, which is not such an unlikely scenario.  We
should handle it instead of just exiting.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>