Fix #393: defer setting GnuTLS CA trust to ssl_open()
As mentioned in #391, setting the CA trust for the OpenSSL backend is
deferred to ssl_open().  This patch defers the GnuTLS CA init in a
similar fashion, allowing a custom user-defined trust file to be set
also when using the GnuTLS backend.

Verified manually with the FreeDNS plugin and three different CA files:

 - /etc/ssl/certs/ca-certificates.crt (everything works, CAFILE1)
 - /etc/ssl/certs/UCA_Global_G2_Root.pem (validation fails, expected)
 - /usr/share/ca-certificates/mozilla/USERTrust_RSA_Certification_Authority.crt (works)

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>