ipv6 dualstack support. the simple way
this patch contains various changes for supporting ipv4 and ipv6 at the
same time without messing up too much code

problems found and solutions in this patch:

problem 1: the ip cache file must hold 2 ip addresses for the same host. for ipv4 and for ipv6

solution:
we now create the cache file not based the host/alias name, but combined
from alias/host and the system/provider name. and we create a alias for
each ipv6 supporting provider which is prefixed by ipv6@
so 2 cache files for each individual update will be created then holding
the correct addresses

problem 2: for most providers 2 requests must be done. one for the ipv4
address and one for the ipv6 address. both are using the same provider
update url. this may cause problems with rate limiting (see easy dns,
where rate limit is one request per 10 minutes.

solution:
currently the force update parameter is global and set to 1 if a update
fails due rate limiting and is set to retry. since we have to create 2
provider entries in config file both will be updated on the same time
over and over again running into a unsuccessfull update for one of both
addresses. this patch changes this to keep the force update parameter
individually for each alias. so a already successfull request will not
be requested again in the forced addr update period.

additional info:
the alias for each ipv6 supporting provider has also several other
advantages. we can now use a single checkip script. the script may then
check for the env parameter INADYN_PROVIDER to check if it was a ipv6 or
ipv4 request. this allows a more cleaner implementation from user side.
to save code most of the aliases are created by simply calling
register_plugin_v6. the function will creates a clone of the plugin
provider and changes the prefix default@ to ipv6@. plugin_unregister
does automaticly release this alias if found

additional fixes in this patch:
1.
use correct new api of zoneedit which is now identical to easydns api.
this brings ipv6 support of course
2.
add ipv6 support to duiadns
3.
use dynsize.php instead of generic.php for easydns (and zoneedit of
course). generic.php returns html code in error responses. dynsite.php
is using plain text which is way more usefull for logging
4.
ipv6 support for gira.de

Signed-off-by: Sebastian Gottschall <s.gottschall@dd-wrt.com>