Fix #421: use Docker :latest tag only for latest In-a-Dyn release
Ever since first adding support for building containers, PR #329,
In-a-Dyn container images have applied the :latest tag wrong.

According to best practices[1], the :latest tag should be applied
to the latest release of a software, not the latest build.

This patch changes that, using GitHub action variables[2] to set up a
list of tags or generate a version for tagging.  The docker builders
take multiple tags separated with comma[3] (',') and for GHCR we simply
use a for-loop.

[1]: https://stevelasker.blog/2018/03/01/docker-tagging-best-practices-for-tagging-and-versioning-docker-images/
[2]: https://docs.github.com/en/actions/learn-github-actions/variables
[3]: https://github.com/docker/build-push-action

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>