Remove duplicate branches in request of plugins
The check for whether the name contains "ipv6" is not necessary for
the plugins "dnshome" and "ipv64", as the code path for both branches is the
same.
Furthermore the check is rather confusing in the context of the
provider named "ipv64.net", as "ipv6" is also contained in the
provider name itself.