Provider ipv64.net: Fix issue with detection whether to use IPv6 or IPv4
The issue was occurring while checking whether the provider name contained
"ipv6" as an indication for whether to use IPv4 or IPv4. The check
confused a definition like "ipv4@ipv64.net" to use IPv6 due to having
"ipv6" in the name.

To address the issue we changed the check to whether the prefix of
the provider name is "ipv6".

While doing so, we factored out this check into a own function taking
a ddns_info_* struct as an argument.